store answers in localStorage, plus a refactor of components dir
