+++ src/components/SignOutButton.tsx
@@ -1,0 +1,20 @@
+import { useAuth } from "~/lib/firebase";
+
+type Props = {};
+
+export const SignOutButton = (props: Props) => {
+  const handleClick = () => {
+    const auth = useAuth();
+    auth.signOut();
+  };
+
+  return (
+    <button
+      onClick={handleClick}
+      type="button"
+      className="btn normal-case"
+    >
+      Sign Out
+    </button>
+  );
+};
