--- src/components/router/Router.tsx
@@ -1,66 +1,0 @@
-import { lazy, Suspense } from 'react';
-import { Outlet, RouteObject, useRoutes, BrowserRouter } from 'react-router-dom';
-
-const Loading = () => <p className="p-4 w-full h-full text-center">Loading...</p>;
-
-const IndexScreen = lazy(() => import('~/components/screens/Index'));
-const QuestionsScreen = lazy(() => import('~/components/screens/Questions'));
-const Page404Screen = lazy(() => import('~/components/screens/404'));
-
-function Layout() {
-  return (
-    <div>
-      <nav className="p-4 flex items-center justify-between">
-        <span>Header</span>
-      </nav>
-      <Outlet />
-    </div>
-  );
-}
-
-export const Router = () => {
-  return (
-    <BrowserRouter>
-      <InnerRouter />
-    </BrowserRouter>
-  );
-};
-
-const InnerRouter = () => {
-  const routes: RouteObject[] = [
-    {
-      path: '/',
-      element: <Layout />,
-      children: [
-        {
-          index: true,
-          element: <IndexScreen />,
-        },
-        {
-          path: '*',
-          element: <Page404Screen />,
-        },
-      ],
-    },
-    {
-      path: '/questions',
-      element: <Layout />,
-      children: [
-        {
-          index: true,
-          element: <QuestionsScreen />,
-        },
-        {
-          path: '*',
-          element: <Page404Screen />,
-        },
-      ],
-    },
-  ];
-  const element = useRoutes(routes);
-  return (
-    <div>
-      <Suspense fallback={<Loading />}>{element}</Suspense>
-    </div>
-  );
-};
