--- src/components/screens/Index.tsx
@@ -1,101 +1,0 @@
-import { Head } from '~/components/shared/Head';
-import JoinGroup from './JoinGroup';
-
-function Index() {
-  
-
-  return (
-    <>
-      <Head title="TOP PAGE" />
-      <div className="container mx-auto h-screen">
-        <JoinGroup />
-      </div>
-      
-        {/*
-
-      <div className="hero min-h-screen">
-        <div className="hero-content">
-
-          <div>
-            <h1 className="text-3xl font-bold">
-              <a className="link link-primary" target="_blank" href="https://vitejs.dev/" rel="noreferrer">
-                Vite
-              </a>{' '}
-              +{' '}
-              <a className="link link-primary" target="_blank" href="https://reactjs.org/" rel="noreferrer">
-                React
-              </a>{' '}
-              +{' '}
-              <a className="link link-primary" target="_blank" href="https://www.typescriptlang.org/" rel="noreferrer">
-                TypeScript
-              </a>{' '}
-              +{' '}
-              <a className="link link-primary" target="_blank" href="https://tailwindcss.com/" rel="noreferrer">
-                TailwindCSS
-              </a>{' '}
-              Starter
-            </h1>
-            <p className="mt-4 text-lg">
-              For fast <b>prototyping</b>. Already set up{' '}
-              <a
-                className="link link-primary"
-                target="_blank"
-                href="https://github.com/firebase/firebase-js-sdk"
-                rel="noreferrer"
-              >
-                Firebase(v9)
-              </a>
-              ,{' '}
-              <a className="link link-primary" target="_blank" href="https://daisyui.com/" rel="noreferrer">
-                daisyUI
-              </a>
-              ,{' '}
-              <a className="link link-primary" target="_blank" href="https://github.com/eslint/eslint" rel="noreferrer">
-                ESLint
-              </a>
-              ,{' '}
-              <a
-                className="link link-primary"
-                target="_blank"
-                href="https://github.com/prettier/prettier"
-                rel="noreferrer"
-              >
-                Prettier
-              </a>
-              .
-            </p>
-            <div className="mt-4 grid gap-2">
-              {state.state === 'UNKNOWN' ? null : state.state === 'SIGNED_OUT' ? <SignInButton /> : <SignOutButton />}
-              <button onClick={() => setIsOpen(true)}>Display Dialog</button>
-            </div>
-          </div>
-        </div>
-      </div>
-      <Dialog
-        className="flex fixed inset-0 z-10 overflow-y-auto"
-        initialFocus={completeButtonRef}
-        open={isOpen}
-        onClose={() => setIsOpen(false)}
-      >
-        <div className="flex items-center justify-center min-h-screen w-screen">
-          <Dialog.Overlay className="fixed inset-0 bg-black opacity-30" />
-          <div className="relative bg-white rounded max-w-120 p-8 mx-auto">
-            <Dialog.Title>Dialog Title</Dialog.Title>
-            <Dialog.Description>Dialog description</Dialog.Description>
-            <button
-              ref={completeButtonRef}
-              type="button"
-              className="inline-flex justify-center px-4 py-2 text-sm font-medium text-blue-900 bg-blue-100 border border-transparent rounded-md hover:bg-blue-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-blue-500"
-              onClick={() => setIsOpen(false)}
-            >
-              Got it, thanks!
-            </button>
-          </div>
-        </div>
-      </Dialog>
-  */}
-    </>
-  );
-}
-
-export default Index;
